Refactored to move fixture function.
